@@ -1,21 +1,28 @@
 {
   "private": true,
   "scripts": {
-    "dev": "next dev",
+    "dev": "next dev --turbopack",
     "build": "next build",
     "start": "next start"
   },
   "dependencies": {
-    "framer-motion": "10.8.4",
-    "next": "latest",
-    "react": "18.2.0",
-    "react-dom": "18.2.0",
-    "react-player": "^2.12.0",
-    "styled-components": "^5.3.6"
+    "babel-plugin-react-compiler": "^19.0.0-beta-27714ef-20250124",
+    "clsx": "^2.1.1",
+    "framer-motion": "^11.17.0",
+    "next": "15.1.4",
+    "react": "^19.0.0",
+    "react-dom": "^19.0.0",
+    "react-player": "^2.12.0"
   },
   "devDependencies": {
-    "@types/node": "18.14.2",
-    "@types/react": "18.0.28",
-    "typescript": "4.9.5"
+    "@eslint/eslintrc": "^3",
+    "@types/node": "^20",
+    "@types/react": "^19",
+    "@types/react-dom": "^19",
+    "eslint": "^9",
+    "eslint-config-next": "15.1.4",
+    "postcss": "^8",
+    "tailwindcss": "^3.4.1",
+    "typescript": "^5"
   }
 }
