@@ -1,29 +1,14 @@
-import { useRouter } from "next/router";
-import { useMemo } from "react";
-import { projects } from "../constants";
-import { Container } from "../styles/styles";
+"use client";
 
+import { projects } from "../constants";
 import ProjectRow from "./ProjectRow";
-import { Grid } from "./Projects.styles";
+import { motion } from "framer-motion";
 
 const Projects = () => {
-  const { query } = useRouter();
-
-  const filteredProjects = useMemo(() => {
-    if (query.filter) {
-      return projects.filter((project) => {
-        if (project.category === query.filter) {
-          return project;
-        }
-      });
-    }
-
-    return projects;
-  }, [query.filter]);
-
   return (
-    <Container style={{ marginBottom: "200px" }} key={query.filter}>
-      <Grid
+    <div className="container mx-auto px-4 mb-48">
+      <motion.div
+        className="flex flex-col rounded-md overflow-hidden"
         initial={{
           y: 40,
           opacity: 0,
@@ -34,13 +19,14 @@ const Projects = () => {
         }}
         transition={{
           duration: 0.5,
+          ease: [0.76, 0, 0.24, 1],
         }}
       >
-        {filteredProjects.map((project, index) => {
-          return <ProjectRow key={index} project={project} index={index} />;
-        })}
-      </Grid>
-    </Container>
+        {projects.map((project, index) => (
+          <ProjectRow key={index} project={project} index={index} />
+        ))}
+      </motion.div>
+    </div>
   );
 };
 
