@@ -1,11 +1,6 @@
-import { Container } from "../styles/styles";
-import {
-  ListItem,
-  Wrapper,
-  ListsContainer,
-  List,
-  Anchor,
-} from "./Footer.styles";
+"use client";
+
+import { motion } from "framer-motion";
 
 const initial = { opacity: 0, y: -10 };
 const whileInView = { opacity: 1, y: 0 };
@@ -13,69 +8,80 @@ const viewport = { once: true };
 
 const Footer = () => {
   return (
-    <Wrapper>
-      <Container>
-        <ListsContainer>
-          <List>
-            <ListItem
-              uppercase
+    <div className="bg-dark flex-1 min-h-[50vh] flex justify-end items-end">
+      <div className="container mx-auto px-4">
+        <div className="grid grid-cols-2 sm:grid-cols-3 gap-8 w-full sm:w-1/2 mb-24">
+          <ul>
+            <motion.li
               initial={initial}
               whileInView={whileInView}
               viewport={viewport}
+              className="text-text-secondary font-[var(--geomanist-font)] text-base leading-relaxed uppercase mb-1.5 last:mb-0"
             >
               GIF produkcija
-            </ListItem>
-            <ListItem
+            </motion.li>
+            <motion.li
               initial={initial}
               whileInView={whileInView}
               viewport={viewport}
               transition={{ delay: 0.1 }}
+              className="text-text-secondary font-[var(--geomanist-font)] text-base leading-relaxed mb-1.5 last:mb-0"
             >
               Radićeva 16
-            </ListItem>
-            <ListItem
+            </motion.li>
+            <motion.li
               initial={initial}
               whileInView={whileInView}
               viewport={viewport}
               transition={{ delay: 0.2 }}
+              className="text-text-secondary font-[var(--geomanist-font)] text-base leading-relaxed mb-1.5 last:mb-0"
             >
               47000 Karlovac
-            </ListItem>
-          </List>
-          <List>
-            <ListItem
+            </motion.li>
+          </ul>
+          <ul>
+            <motion.li
               initial={initial}
               whileInView={whileInView}
               viewport={viewport}
+              className="text-text-secondary font-[var(--geomanist-font)] text-base leading-relaxed mb-1.5 last:mb-0"
             >
               kontakt@gif.hr
-            </ListItem>
-            <ListItem
+            </motion.li>
+            <motion.li
               initial={initial}
               whileInView={whileInView}
               viewport={viewport}
               transition={{ delay: 0.1 }}
+              className="text-text-secondary font-[var(--geomanist-font)] text-base leading-relaxed mb-1.5 last:mb-0"
             >
-              <Anchor href="tel:+385977929943">+385 (97) 792 9943</Anchor>
-            </ListItem>
-          </List>
-          <List>
-            <ListItem
+              <a
+                href="tel:+385977929943"
+                className="text-text-secondary font-[var(--geomanist-font)] text-base leading-relaxed no-underline"
+              >
+                +385 (97) 792 9943
+              </a>
+            </motion.li>
+          </ul>
+          <ul>
+            <motion.li
               initial={initial}
               whileInView={whileInView}
               viewport={viewport}
+              className="text-text-secondary font-[var(--geomanist-font)] text-base leading-relaxed mb-1.5 last:mb-0"
             >
-              <Anchor
+              <a
                 href="https://www.facebook.com/gifprodukcija"
                 target="_blank"
+                className="text-text-secondary font-[var(--geomanist-font)] text-base leading-relaxed no-underline"
               >
                 Facebook
-              </Anchor>
-            </ListItem>
-          </List>
-        </ListsContainer>
-      </Container>
-    </Wrapper>
+              </a>
+            </motion.li>
+          </ul>
+        </div>
+      </div>
+    </div>
   );
 };
 
