@@ -1,9 +1,8 @@
-import { useScroll, useTransform } from "framer-motion";
-import Image from "next/image";
-import { useUI } from "../context/UIProvider";
-import { Container } from "../styles/styles";
+"use client";
 
-import { Button, NavBar, Wrapper } from "./Navigation.styles";
+import { useScroll, useTransform, motion } from "framer-motion";
+import Image from "next/image";
+import { useUI } from "@/context/UIProvider";
 
 const Navigation = () => {
   const { setApplyOverlay } = useUI();
@@ -14,24 +13,21 @@ const Navigation = () => {
   });
 
   return (
-    <NavBar
-      initial={{
-        opacity: 0,
-      }}
-      animate={{
-        opacity: 1,
-      }}
-      transition={{
-        duration: 0.5,
-      }}
+    <motion.nav
+      initial={{ opacity: 0 }}
+      animate={{ opacity: 1 }}
+      transition={{ duration: 0.5 }}
+      className="sticky top-0 z-20 mb-32"
     >
-      <Container>
-        <Wrapper>
-          <Image alt="logo" src="/images/logo.svg" width={166} height={32} />
-          <Button>Kontakt</Button>
-        </Wrapper>
-      </Container>
-    </NavBar>
+      <div className="container mx-auto px-4">
+        <div className="flex justify-between items-center py-6">
+          <Image alt="logo" src="/images/logo.svg" width={32} height={32} />
+          <button className="group relative inline-flex h-12 items-center justify-center overflow-hidden rounded-md border border-neutral-200 px-6 text-white bg-black transition-all duration-100 [box-shadow:5px_5px_rgb(255_255_255)] active:translate-x-[3px] active:translate-y-[3px] active:[box-shadow:0px_0px_rgb(255_255_255)]">
+            Contact
+          </button>
+        </div>
+      </div>
+    </motion.nav>
   );
 };
 
